add fetch command, and less-hackiness
allows you to quickly fetch a single table with your configuration
info. this way you can quickly grab a table if you need to do a quick
data check on it, or something with a particular table.

i also use the "re" library in nodejs so we can implement
a less hacky backoff function just felt why not clean
up the techincal debt now. instead of fixing it later.

also bumped the version to v0.2.0 so it can be updated

Change-Id: Id63658a693505c9f7f773ba799fb25fd8f282dd3