Replace Sync implementation
This replaces the sync implementation with a new version
that uses a new API. The `api/file/sync` endpoint does most of the same
work the CLI previously did of getting the current view of which
files should exist. This greatly simplify the CLI code and should also
be much much simpler for other languages to implement

This implementation also has the benefit of being restartable and does
away with the StateStore.

Other changes:
- Docs are updated to reflect new API
- Config file is now written out

Change-Id: Ic403a66876ea1469d1a75eec35fbfe39775d9069
Reviewed-on: https://gerrit.instructure.com/76841
Tested-by: Jenkins
Reviewed-by: Matt Smith <msmith@instructure.com>
Product-Review: Addison Higham <ahigham@instructure.com>
QA-Review: Addison Higham <ahigham@instructure.com>