allow json output for the list command
this adds in the an option for lists that allows one to specify
json output (e.g. to parse the output from a machine standpoint).

test-plan:
  - simply install the CLI
  - run: `canvasDataCli list -c config.js` and see the normal output.
  - run: `canvasDataCli list -c config.js -j` and see json output.

If you want to make the JSON output pretty to verify you can do
that with JQ. however this is mainly added in for machine parsing
I don't expect a human to ever use this without JQ.

Change-Id: I39f5146241277a32f1f7ecdb3ced8a93a19c8139
Reviewed-on: https://gerrit.instructure.com/107685
Reviewed-by: Addison Higham <ahigham@instructure.com>
Product-Review: Addison Higham <ahigham@instructure.com>
QA-Review: Addison Higham <ahigham@instructure.com>
Tested-by: Jenkins